+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.1.35
 - new UI for Collections page
 - updates for wallet options
+- updates for connect wallet screen (smart wallet integration)
 
 ## 2.1.34
 - React, Wagmi, web3Modal update